+++ test/com/amplitude/api/AmplitudeClientTest.java
@@ -56,7 +56,7 @@ public class AmplitudeClientTest extends BaseTest {
 
     @Test
     public void testSetUserId() {
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         ShadowLooper looper = Shadows.shadowOf(amplitude.logThread.getLooper());
         String userId = "user_id";
         amplitude.setUserId(userId);
@@ -100,7 +100,7 @@ public class AmplitudeClientTest extends BaseTest {
 
     @Test
     public void testSetDeviceId() {
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         ShadowLooper looper = Shadows.shadowOf(amplitude.logThread.getLooper());
         assertNull(amplitude.getDeviceId());
         looper.runToEndOfTasks();
@@ -238,7 +238,7 @@ public class AmplitudeClientTest extends BaseTest {
         ShadowLooper looper = Shadows.shadowOf(amplitude.logThread.getLooper());
 
         assertNull(amplitude.getDeviceId());
-        DatabaseHelper.getDatabaseHelper(context).insertOrReplaceKeyValue(
+        DatabaseHelper.getDatabaseHelper(context, apiKey).insertOrReplaceKeyValue(
                 AmplitudeClient.DEVICE_ID_KEY,
                 deviceId
         );
@@ -259,7 +259,7 @@ public class AmplitudeClientTest extends BaseTest {
         looper.getScheduler().advanceToLastPostedRunnable();
         String deviceId = amplitude.getDeviceId();
         assertTrue(deviceId.endsWith("R"));
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         assertEquals(
                 deviceId,
                 dbHelper.getValue(AmplitudeClient.DEVICE_ID_KEY)
@@ -275,7 +275,7 @@ public class AmplitudeClientTest extends BaseTest {
         assertEquals(37, amplitude.getDeviceId().length());
         String deviceId = amplitude.getDeviceId();
         assertTrue(deviceId.endsWith("R"));
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         assertEquals(
                 deviceId,
                 dbHelper.getValue(AmplitudeClient.DEVICE_ID_KEY)
@@ -287,7 +287,7 @@ public class AmplitudeClientTest extends BaseTest {
         ShadowLooper looper = Shadows.shadowOf(amplitude.logThread.getLooper());
         ShadowLooper httplooper = Shadows.shadowOf(amplitude.httpThread.getLooper());
 
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         assertFalse(amplitude.isOptedOut());
         assertEquals((long) dbHelper.getLongValue(AmplitudeClient.OPT_OUT_KEY), 0L);
 
@@ -361,7 +361,7 @@ public class AmplitudeClientTest extends BaseTest {
         assertTrue(compareJSONObjects(userProperties.getJSONObject(Constants.AMP_OP_SET), expected));
 
         // verify db state
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         assertNull(dbHelper.getValue(AmplitudeClient.USER_ID_KEY));
         assertEquals((long)dbHelper.getLongValue(AmplitudeClient.LAST_IDENTIFY_ID_KEY), 1L);
         assertEquals((long)dbHelper.getLongValue(AmplitudeClient.LAST_EVENT_ID_KEY), -1L);
@@ -606,7 +606,7 @@ public class AmplitudeClientTest extends BaseTest {
         assertEquals(getUnsentIdentifyCount(), 0);
 
         // verify db state
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         assertNull(dbHelper.getValue(AmplitudeClient.USER_ID_KEY));
         assertEquals((long) dbHelper.getLongValue(AmplitudeClient.LAST_IDENTIFY_ID_KEY), 3L);
         assertEquals((long) dbHelper.getLongValue(AmplitudeClient.LAST_EVENT_ID_KEY), 4L);
@@ -638,7 +638,7 @@ public class AmplitudeClientTest extends BaseTest {
         event.remove("sequence_number");
         event.remove("event_id");
         // delete event from db and reinsert modified event
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         dbHelper.removeEvent(1);
         dbHelper.addEvent(event.toString());
         amplitude.uploadingCurrently.set(false);
@@ -970,7 +970,7 @@ public class AmplitudeClientTest extends BaseTest {
         clock.setTimestamps(timestamps);
         Robolectric.getForegroundThreadScheduler().advanceTo(1);
 
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         ShadowLooper looper = Shadows.shadowOf(amplitude.logThread.getLooper());
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 0);
@@ -1147,7 +1147,7 @@ public class AmplitudeClientTest extends BaseTest {
 
     @Test
     public void testAutoIncrementSequenceNumber() {
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         int limit = 10;
         for (int i = 0; i < limit; i++) {
             assertEquals(amplitude.getNextSequenceNumber(), i+1);
@@ -1184,7 +1184,7 @@ public class AmplitudeClientTest extends BaseTest {
         clock.setTimestamps(timestamps);
         Robolectric.getForegroundThreadScheduler().advanceTo(1);
 
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         int eventMaxCount = 3;
         ShadowLooper looper = Shadows.shadowOf(amplitude.logThread.getLooper());
         amplitude.setEventMaxCount(eventMaxCount).setOffline(true);
@@ -1220,7 +1220,7 @@ public class AmplitudeClientTest extends BaseTest {
 
     @Test
     public void testTruncateEventsQueues() {
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         int eventMaxCount = 50;
         assertTrue(eventMaxCount > Constants.EVENT_REMOVE_BATCH_SIZE);
         ShadowLooper looper = Shadows.shadowOf(amplitude.logThread.getLooper());
@@ -1239,7 +1239,7 @@ public class AmplitudeClientTest extends BaseTest {
 
     @Test
     public void testTruncateEventsQueuesWithOneEvent() {
-        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context, apiKey);
         int eventMaxCount = 1;
         ShadowLooper looper = Shadows.shadowOf(amplitude.logThread.getLooper());
         amplitude.setEventMaxCount(eventMaxCount).setOffline(true);