+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Migrate Sqlite database file to new filename with apiKey.
+
 ## 2.9.0 (July 07, 2016)
 
 * Add automatic flushing of unsent events on app close/minimize (through the Activity Lifecycle `onPause` callback). This only works if you call `Amplitude.getInstance().enableForegroundTracking(getApplication());`, which is recommended in the README by default for Setup. To disable you can call `Amplitude.getInstance().setFlushEventsOnClose(false);`